+++ lib/surface/type_handler/hook.ex
@@ -32,7 +32,7 @@ defmodule Surface.TypeHandler.Hook do
     ctx = Surface.AST.Meta.quoted_caller_context(meta)
 
     quoted_expr =
-      quote generated: true do
+      quote do
         Surface.TypeHandler.expr_to_value!(
           unquote(type),
           unquote(name),