+++ test/surface/live_component_test.exs
@@ -54,20 +54,20 @@ defmodule Surface.LiveComponentTest do
   defmodule InfoProvider do
     use Surface.Component
 
-    slot default, args: [:info]
+    slot default, arg: %{info: :string}
 
     def render(assigns) do
       info = "Hi there!"
 
       ~F"""
       <div>
-        <#slot :args={info: info}/>
+        <#slot {@default, info: info}/>
       </div>
       """
     end
   end
 
-  defmodule InfoProviderWithoutSlotArgs do
+  defmodule InfoProviderWithoutSlotArg do
     use Surface.Component
 
     slot default
@@ -105,13 +105,13 @@ defmodule Surface.LiveComponentTest do
     end
   end
 
-  test "render content without slot args" do
+  test "render content without slot arg" do
     html =
       render_surface do
         ~F"""
-        <InfoProviderWithoutSlotArgs>
+        <InfoProviderWithoutSlotArg>
           <span>Hi there!</span>
-        </InfoProviderWithoutSlotArgs>
+        </InfoProviderWithoutSlotArg>
         """
       end
 
@@ -122,7 +122,7 @@ defmodule Surface.LiveComponentTest do
            """
   end
 
-  test "render content with slot args" do
+  test "render content with slot arg" do
     html =
       render_surface do
         ~F"""